social_media_agent/requirements.txt
fastapi
uvicorn
youtube-transcript-api
yt-dlp
requests
faster-whisper
torch
python-dotenv